Renamed styled-components folder
