+++ src/components/shared-components/StyledHr.jsx
